+++ tradeoffer.go
@@ -64,7 +64,7 @@ type EconItem struct {
 	InstanceID uint64 `json:"instanceid,string,omitempty"`
 	ClassID    uint64 `json:"classid,string,omitempty"`
 	AppID      uint32 `json:"appid,string"`
-	ContextID  uint16 `json:"contextid,string"`
+	ContextID  uint64 `json:"contextid,string"`
 	Amount     uint16 `json:"amount,string"`
 	Name       string `json:"name,string"` // Will be used for item descriptions, do *not* remove
 	Missing    bool   `json:"missing,omitempty"`
